requirements.txt
discord
discord.py
python-dotenv
pillow
